k8sutil: fix UpdateNodeRetry to handle conflicts
Prior to this change it claimed to handle conflicts, but because it
didn't 'Get' the new node, it would continue to try and submit the out
of date update which caused a problem in the first place.

This updates the logic, adds a test, and exports the function (since
it's planned for broader use in an upcoming change).