operator: account for reboots already underway
This makes the operator's logic more restart-safe.

This also moves the 1 hour sleep/wait pair previously used into
eventloop polling.
This is mildly less efficient, but it's closer to what we want
eventually.
We can regain the desirable properties of `watch`-ing in the future by
watching certain node annotations.