operator: add leader-election
This makes the operator behavior safer. Prior to this, it would be
possible for two operators to run and stomp on each other in a few
circumstances.

With this change, any write actions are only done after checking that
this operator is still the leader.

This makes it quite unlikely that any conflict occurs unless there's a
large amount of clock skew or a write operation takes over 20 seconds to
complete.