agent: moved getting first ue status to ReceiveStatuses func
Generating a status and putting it in our channel would deadlock if there was
already a ue status update waiting for us, which happens when a signal comes
along between the time we subscribe in New() and the time we actually start
consuming status updates. Instead, make the ReceiveStatuses function get the
current status for us and the problem goes away.