pkg/operator: added pre- and post-reboot actions
This commit adds pre- and post-reboot actions.

With these changes, you will be able to configure actions which get run before a
reboot is initiated and after one is completed, effectively gating it on either
side. The actions taken are entirely arbitrary, and software that implements
them only needs to be able to manipulate kubernetes annotations and read labels.

The changes ended up being a little more of an overhaul then I originally
intended, but I'm happy with the outcome.